// src/app/features/arquivo/interfaces/create-mark-up.dto.ts
export class CreateMarkUpDto {


  arquivo_comentario_descricao?: string;

  arquivo_id?: string;

  arquivo_comentario_id_pai?: number;

  arquivo_comentario_nivel?: number;

  usuario_id?: number;

  arquivo_comentario_markup_selecao?: number;

  arquivo_comentario_markup_link_s3?: string;

  arquivo_comentario_markup_descricao?: string;
}