// src/app/core/models/usuario.ts
export class Usuario {
  usuario_id:       number | undefined;
  usuario_nome:     string = "";
  usuario_email:    string = "";
  usuario_cpf:      string = "";
  usuario_cnpj:     string = "";
  usuario_endereco: string = "";
  usuario_status:   number | undefined;
  usuario_cargo:    string = "";
  usuario_tipo:     number | undefined;
  empresa_id:       number | undefined;
}